requirements.txt
langchain
python-dotenv
ipykernel
langsmith
langchain_community
langchain_core
streamlit
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
duckdb
pandas
openai
langchain-groq
torch

